Extract @Setup method in base test class
